refactor : ApiUtil 패키지 이동
