+++ seekgu/src/main/java/com/seekgu/utils/ApiUtil.java
@@ -1,4 +1,4 @@
-package com.seekgu;
+package com.seekgu.utils;
 
 public class ApiUtil {
